refactor: split schema
