+++ src/schema/2.0/items-object/array-items-object.ts
@@ -1,0 +1,8 @@
+import { BaseItemsObject } from './base-items-object';
+import { Option } from 'fp-ts/lib/Option';
+import { ItemsObject } from './items-object';
+
+export interface ArrayItemsObject extends BaseItemsObject {
+	readonly type: 'array';
+	readonly items: Option<ItemsObject[]>;
+}
