+++ src/schema/2.0/items-object/boolean-items-object.ts
@@ -1,0 +1,14 @@
+import * as t from 'io-ts';
+import { BaseItemsObject, BaseItemsObjectProps } from './base-items-object';
+
+export interface BooleanItemsObject extends BaseItemsObject {
+	readonly type: 'boolean';
+}
+
+export const BooleanItemsObject = t.type(
+	{
+		...BaseItemsObjectProps,
+		type: t.literal('boolean'),
+	},
+	'BooleanItemsObject',
+);
