+++ src/schema/2.0/items-object/non-array-items-object.ts
@@ -1,0 +1,12 @@
+import { StringItemsObject } from './string-items-object';
+import { NumberItemsObject } from './number-items-object';
+import { IntegerItemsObject } from './integer-items-object';
+import { BooleanItemsObject } from './boolean-items-object';
+import * as t from 'io-ts';
+
+export type NonArrayItemsObject = StringItemsObject | NumberItemsObject | IntegerItemsObject | BooleanItemsObject;
+
+export const NonArrayItemsObject = t.union(
+	[StringItemsObject, NumberItemsObject, IntegerItemsObject, BooleanItemsObject],
+	'NonArrayItemsObject',
+);
