+++ src/schema/2.0/items-object/string-items-object.ts
@@ -1,0 +1,14 @@
+import { BaseItemsObject, BaseItemsObjectProps } from './base-items-object';
+import * as t from 'io-ts';
+
+export interface StringItemsObject extends BaseItemsObject {
+	readonly type: 'string';
+}
+
+export const StringItemsObject = t.type(
+	{
+		...BaseItemsObjectProps,
+		type: t.literal('string'),
+	},
+	'StringItemsObject',
+);
