+++ src/schema/2.0/parameter-object/parameter-object.ts
@@ -1,4 +1,4 @@
-import { PathParameterObject } from './path-parameter-object';
+import { PathParameterObject } from './path-parameter-object/path-parameter-object';
 import { QueryParameterObject } from './query-parameter-object/query-parameter-object';
 import { HeaderParameterObject } from './header-parameter-object';
 import { FormDataParameterObject } from './form-data-parameter-object';
