+++ src/schema/2.0/parameter-object/path-parameter-object/array-parameter-object.ts
@@ -1,0 +1,17 @@
+import { BasePathParameterObjectProps } from './base-parameter-object';
+import * as t from 'io-ts';
+import { NonArrayItemsObject } from '../../items-object/non-array-items-object';
+
+export interface ArrayPathParameterObject extends BasePathParameterObjectProps {
+	readonly type: 'array';
+	readonly items: NonArrayItemsObject;
+}
+
+export const ArrayPathParameterObject = t.type(
+	{
+		...BasePathParameterObjectProps,
+		type: t.literal('array'),
+		items: NonArrayItemsObject,
+	},
+	'ArrayPathParameterObject',
+);
