+++ src/schema/2.0/parameter-object/path-parameter-object/integer-parameter-object.ts
@@ -1,0 +1,14 @@
+import * as t from 'io-ts';
+import { BasePathParameterObjectProps } from './base-parameter-object';
+
+export interface IntegerPathParameterObject extends BasePathParameterObjectProps {
+	readonly type: 'integer';
+}
+
+export const IntegerPathParameterObject = t.type(
+	{
+		...BasePathParameterObjectProps,
+		type: t.literal('integer'),
+	},
+	'IntegerPathParameterObject',
+);
