+++ src/schema/2.0/parameter-object/path-parameter-object/number-parameter-object.ts
@@ -1,0 +1,14 @@
+import { BasePathParameterObjectProps } from './base-parameter-object';
+import * as t from 'io-ts';
+
+export interface NumberPathParameterObject extends BasePathParameterObjectProps {
+	readonly type: 'number';
+}
+
+export const NumberPathParameterObject = t.type(
+	{
+		...BasePathParameterObjectProps,
+		type: t.literal('number'),
+	},
+	'NumberPathParameterObject',
+);
