+++ src/schema/2.0/parameter-object/path-parameter-object/string-parameter-object.ts
@@ -1,0 +1,14 @@
+import * as t from 'io-ts';
+import { BasePathParameterObjectProps } from './base-parameter-object';
+
+export interface StringPathParameterObject extends BasePathParameterObjectProps {
+	readonly type: 'string';
+}
+
+export const StringPathParameterObject = t.type(
+	{
+		...BasePathParameterObjectProps,
+		type: t.literal('string'),
+	},
+	'StringPathParameterObject',
+);
